Update: Use prettier 2.0 and adopt their styles completely
Where before we were using both prettier and eslint for styles. This update adopts prettier more, uses the recommended prettier rules and drops our overrides. This should 1) make it easier to setup (no need for local prettierrc files) 2) solve problems with different IDE setups and conflicts between eslint and prettier and 3) fix a bunch of formatting issues via the prettier upgrade.

When updating to this version, any .prettierrc file in the repository with config overrides should be deleted or there will be conflicts.

Note also that your local `prettier` should be updated to ~2.0.0 - and the same goes for any IDE extensions you might have installed.